python-app/requirements.txt
customtkinter==5.2.0
darkdetect==0.8.0
flatbuffers==23.5.26
humanfriendly==10.0
mpmath==1.3.0
numpy==1.25.2
onnxruntime==1.15.1
Pillow==10.0.0
piper-phonemize==1.1.0
piper-tts==1.2.0
protobuf==4.24.1
PyAudio==0.2.13
pyparsing==3.1.1
